refactor(netcdf): move create_from_array + helpers into engines/variables.py (STR-1)
Relocate the create_from_array constructor and its private helpers
_resolve_extra_dims / _create_netcdf_from_array out of netcdf.py into
module-level functions in engines/variables.py (issue #615).
NetCDF.create_from_array becomes a thin classmethod façade delegating to
the engine function; it still always returns a Container, so behaviour
and return type are unchanged.

The engine function lazily imports Container / NetCDF to break the
netcdf.py <-> engines.variables cycle, reaching the shared static
dimension helpers (get_x_lon_dimension_array / _create_dimension) through
the class. Repoint the two _create_netcdf_from_array validation unit
tests at the helper's new module location (it is private, not public).
