+++ src/pyramids/netcdf/netcdf.py
@@ -34,6 +34,7 @@ from pyramids.netcdf._lazy import apply_unpack, build_lazy_array
 from pyramids.netcdf._mfdataset import open_mfdataset
 from pyramids.netcdf._plot import NetCDFPlot
 from pyramids.netcdf.engines import interop as _interop
+from pyramids.netcdf.engines import variables as _variables
 from pyramids.netcdf.engines.interop import Interop
 from pyramids.netcdf.engines.variables import Variables
 from pyramids.netcdf.cf import (
@@ -4746,420 +4747,14 @@ class NetCDF(Dataset):
         return dim
 
     @classmethod
-    def create_from_array(  # type: ignore[override]
-        cls,
-        arr: np.ndarray,
-        geo: tuple[float, float, float, float, float, float] | None = None,
-        epsg: str | int = 4326,
-        no_data_value: Any | list = DEFAULT_NO_DATA_VALUE,
-        path: str | Path | None = None,
-        variable_name: str | None = None,
-        extra_dim_name: str = "time",
-        extra_dim_values: list | None = None,
-        extra_dims: list[tuple[str, list | None]] | None = None,
-        top_left_corner: tuple[float, float] | None = None,
-        cell_size: int | float | None = None,
-        chunk_sizes: tuple | list | None = None,
-        compression: str | None = None,
-        compression_level: int | None = None,
-        title: str | None = None,
-        institution: str | None = None,
-        source: str | None = None,
-        history: str | None = None,
-    ) -> NetCDF:
-        """Create a NetCDF dataset from a NumPy array and geotransform.
+    def create_from_array(cls, *args, **kwargs) -> "NetCDF":
+        """Facade — :func:`create_from_array <pyramids.netcdf.engines.variables.create_from_array>`.
 
-        For 3-D arrays the first axis is treated as a non-spatial
-        dimension (time, level, depth, etc.) whose name and coordinate
-        values are controlled by `extra_dim_name` and
-        `extra_dim_values`.
-
-        For 4-D+ arrays — e.g. `(time, level, lat, lon)` — pass
-        `extra_dims=[("time", time_values), ("pressure_level", level_values)]`
-        in storage order. Every non-spatial dimension is then
-        materialised on the resulting NetCDF, preserving the full
-        layout. `extra_dims` and the legacy single-dim params
-        (`extra_dim_name` / `extra_dim_values`) are mutually exclusive.
-
-        The driver is inferred from `path`: if `path` is `None`
-        the dataset is created in memory (MEM driver); if a path is
-        provided the netCDF driver writes to disk.
-
-        Args:
-            arr: 2-D `(rows, cols)`, 3-D `(extra_dim, rows, cols)`, or
-                4-D+ `(d_0, ..., d_{n-1}, rows, cols)` NumPy array.
-            geo: Geotransform tuple `(x_min, pixel_size, rotation,
-                y_max, rotation, pixel_size)`.
-            epsg: EPSG code for the spatial reference.
-                Defaults to 4326.
-            no_data_value: Sentinel value for cells outside the
-                domain. Defaults to DEFAULT_NO_DATA_VALUE.
-            path: Output file path. If `None`, the dataset is
-                created in memory. Defaults to None.
-            variable_name: Name of the data variable in the NetCDF
-                file. Defaults to `"data"`.
-            extra_dim_name: Legacy single-dim path. Name of the
-                non-spatial dimension for 3-D arrays (e.g. `"time"`,
-                `"level"`, `"depth"`). Ignored for 2-D arrays.
-                Mutually exclusive with `extra_dims`. Defaults to
-                `"time"`.
-            extra_dim_values: Legacy single-dim path. Coordinate values
-                for the non-spatial dimension. Must have length
-                `arr.shape[0]` for 3-D arrays. Mutually exclusive with
-                `extra_dims`. Defaults to `[0, 1, 2,..., N-1]`.
-            extra_dims: Multi-dim path. Ordered list of
-                `(dim_name, values)` pairs describing every non-spatial
-                dimension in storage order. `len(extra_dims)` must
-                equal `arr.ndim - 2`. Each `values` is either a list of
-                length `arr.shape[i]` or `None` (use integer indices
-                `[0, 1, ..., size - 1]`). Mutually exclusive with
-                `extra_dim_name` / `extra_dim_values`.
-            top_left_corner: `(x, y)` of the top-left corner. Used
-                with `cell_size` to build `geo` when `geo` is
-                not provided. Defaults to None.
-            cell_size: Pixel size. Used with `top_left_corner` to
-                build `geo`. Defaults to None.
-            chunk_sizes: Chunk sizes for the data variable as a tuple
-                matching the array dimensions (e.g. `(1, 256, 256)`
-                for 3-D). Only effective when writing to disk.
-                Defaults to None (GDAL default chunking).
-            compression: Compression algorithm name (`"DEFLATE"`,
-                `"ZSTD"`, etc.). Only effective when writing to
-                disk. Defaults to None (no compression).
-            compression_level: Compression level (e.g. 1-9 for
-                DEFLATE). Defaults to None (GDAL default).
-            title: CF global attribute `title`. Short
-                description of the dataset. Defaults to None.
-            institution: CF global attribute `institution`.
-                Where the data was produced. Defaults to None.
-            source: CF global attribute `source`. How the
-                data was produced. Defaults to None.
-            history: CF global attribute `history`. Audit
-                trail of processing steps. Defaults to None.
-
-        Returns:
-            NetCDF: The newly created NetCDF dataset.
+        Builds a new :class:`Container` from a NumPy array; the full signature and
+        contract live in the engine function. ``create_from_array`` always returns a
+        ``Container`` regardless of the subtype the classmethod is invoked on.
         """
-        if geo is None and top_left_corner is not None and cell_size is not None:
-            geo = (
-                top_left_corner[0],
-                cell_size,
-                0,
-                top_left_corner[1],
-                0,
-                -cell_size,
-            )
-        if geo is None:
-            raise ValueError(
-                "Either 'geo' or both 'top_left_corner' and "
-                "'cell_size' must be provided."
-            )
-
-        rows = int(arr.shape[-2]) if arr.ndim >= 2 else 0
-        cols = int(arr.shape[-1]) if arr.ndim >= 2 else 0
-
-        # Reconcile the legacy single-dim params with the new
-        # `extra_dims` list-of-pairs API. Result is a normalised list
-        # of (name, values) pairs whose length equals
-        # `max(arr.ndim - 2, 0)`.
-        resolved_extra_dims = cls._resolve_extra_dims(
-            arr=arr,
-            extra_dim_name=extra_dim_name,
-            extra_dim_values=extra_dim_values,
-            extra_dims=extra_dims,
-        )
-
-        if arr.ndim == 3:
-            ClassicDimensionInfo(
-                name=resolved_extra_dims[0][0],
-                size=arr.shape[0],
-                values=resolved_extra_dims[0][1],
-            )
-
-        if variable_name is None:
-            variable_name = "data"
-
-        dst_ds = cls._create_netcdf_from_array(
-            arr,
-            variable_name,
-            cols,
-            rows,
-            resolved_extra_dims,
-            geo,
-            epsg,
-            no_data_value,
-            path=path,
-            chunk_sizes=chunk_sizes,
-            compression=compression,
-            compression_level=compression_level,
-            title=title,
-            institution=institution,
-            source=source,
-            history=history,
-        )
-        result = Container(dst_ds)
-
-        return result
-
-    @staticmethod
-    def _resolve_extra_dims(
-        arr: np.ndarray,
-        extra_dim_name: str,
-        extra_dim_values: list | None,
-        extra_dims: list[tuple[str, list | None]] | None,
-    ) -> list[tuple[str, list]]:
-        """Normalise the legacy + new extra-dim API into a single list.
-
-        Returns an ordered list of `(dim_name, values)` pairs whose
-        length equals `max(arr.ndim - 2, 0)`. Each `values` entry is a
-        concrete Python list (never `None` — defaults are filled with
-        integer indices `[0, 1, ..., size - 1]`).
-
-        Args:
-            arr: The data array; only its `ndim` and `shape` are read.
-            extra_dim_name: Legacy single-dim name (caller-default
-                `"time"`).
-            extra_dim_values: Legacy single-dim values, or `None`.
-            extra_dims: New multi-dim list of `(name, values)` pairs,
-                or `None` for the legacy path.
-
-        Returns:
-            list[tuple[str, list]]: Normalised dim specs.
-
-        Raises:
-            ValueError: If `extra_dims` is supplied alongside
-                `extra_dim_values`; if `extra_dims` length doesn't
-                match `arr.ndim - 2`; or if any per-dim `values`
-                length doesn't match the corresponding `arr.shape[i]`.
-        """
-        expected = max(arr.ndim - 2, 0)
-        if extra_dims is not None:
-            if extra_dim_values is not None:
-                raise ValueError(
-                    "extra_dims and extra_dim_values are mutually "
-                    "exclusive. Use one or the other."
-                )
-            if len(extra_dims) != expected:
-                raise ValueError(
-                    f"extra_dims must have {expected} entries for a "
-                    f"{arr.ndim}-D array, got {len(extra_dims)}."
-                )
-            resolved: list[tuple[str, list]] = []
-            for i, entry in enumerate(extra_dims):
-                name, values = entry
-                if values is None:
-                    values = list(range(int(arr.shape[i])))
-                elif len(values) != int(arr.shape[i]):
-                    raise ValueError(
-                        f"extra_dims[{i}] values length {len(values)} "
-                        f"does not match arr.shape[{i}]={arr.shape[i]}."
-                    )
-                else:
-                    values = list(values)
-                resolved.append((name, values))
-            return resolved
-        if expected == 0:
-            return []
-        if expected == 1:
-            values = (
-                list(extra_dim_values)
-                if extra_dim_values is not None
-                else list(range(int(arr.shape[0])))
-            )
-            return [(extra_dim_name, values)]
-        # 4-D+ array with no `extra_dims` and no legacy values: fall
-        # back to anonymous dim names and integer indices so the array
-        # can still be written.
-        return [(f"dim_{i}", list(range(int(arr.shape[i])))) for i in range(expected)]
-
-    @staticmethod
-    def _create_netcdf_from_array(
-        arr: np.ndarray,
-        variable_name: str,
-        cols: int,
-        rows: int,
-        extra_dims: list[tuple[str, list]] | None = None,
-        geo: tuple[float, float, float, float, float, float] | None = None,
-        epsg: str | int | None = None,
-        no_data_value: Any | list = DEFAULT_NO_DATA_VALUE,
-        path: str | Path | None = None,
-        chunk_sizes: tuple | list | None = None,
-        compression: str | None = None,
-        compression_level: int | None = None,
-        title: str | None = None,
-        institution: str | None = None,
-        source: str | None = None,
-        history: str | None = None,
-    ) -> gdal.Dataset:
-        """Build a multidimensional GDAL dataset from an array.
-
-        The driver is inferred from `path`: `None` -> MEM (in-memory),
-        otherwise the netCDF driver writes to disk.
-
-        Args:
-            arr: 2-D `(rows, cols)`, 3-D `(extra_dim, rows, cols)`, or
-                4-D+ `(d_0, ..., d_{n-1}, rows, cols)` NumPy array.
-            variable_name: Name of the data variable.
-            cols: Number of columns.
-            rows: Number of rows.
-            extra_dims: Ordered list of `(dim_name, values)` pairs for
-                every non-spatial dimension. Length matches
-                `arr.ndim - 2`. Empty list for 2-D arrays. Pre-resolved
-                by `_resolve_extra_dims` so each `values` entry is a
-                concrete list.
-            geo: Geotransform tuple. Defaults to None.
-            epsg: EPSG code. Defaults to None.
-            no_data_value: No-data sentinel. Defaults to
-                DEFAULT_NO_DATA_VALUE.
-            path: Output file path. If None, created in memory.
-                Defaults to None.
-            chunk_sizes: Chunk sizes for the variable. Defaults to
-                None.
-            compression: Compression algorithm. Defaults to None.
-            compression_level: Compression level. Defaults to None.
-            title: CF global attribute `title`. Defaults to None.
-            institution: CF global attribute `institution`.
-                Defaults to None.
-            source: CF global attribute `source`.
-                Defaults to None.
-            history: CF global attribute `history`.
-                Defaults to None.
-
-        Returns:
-            gdal.Dataset: The created multidimensional GDAL dataset.
-        """
-        if variable_name is None:
-            raise ValueError("Variable_name cannot be None")
-        if geo is None:
-            raise ValueError("geo cannot be None")
-
-        if extra_dims is None:
-            extra_dims = []
-        dtype = gdal.ExtendedDataType.Create(numpy_to_gdal_dtype(arr))
-        x_dim_values = NetCDF.get_x_lon_dimension_array(geo[0], geo[1], cols)
-        # Y/lat pixel height comes from geo[5] (negative), not geo[1] — using the X cell here would
-        # square a non-square grid (e.g. 2° lon, 1° lat). Pass the positive height abs(geo[5]).
-        y_dim_values = NetCDF.get_y_lat_dimension_array(geo[3], abs(geo[5]), rows)
-
-        if path is not None:
-            driver_type = "netCDF"
-        else:
-            driver_type = "MEM"
-            path = "netcdf"
-
-        src = gdal.GetDriverByName(driver_type).CreateMultiDimensional(str(path))
-        rg = src.GetRootGroup()
-
-        # Set CF global attributes on root group
-        cf_global = {"Conventions": "CF-1.8"}
-        if title is not None:
-            cf_global["title"] = title
-        if institution is not None:
-            cf_global["institution"] = institution
-        if source is not None:
-            cf_global["source"] = source
-        if history is not None:
-            cf_global["history"] = history
-        write_global_attributes(rg, cf_global)
-
-        # Build creation options for chunking and compression
-        create_options = []
-        if chunk_sizes is not None:
-            create_options.append(f"BLOCKSIZE={','.join(str(s) for s in chunk_sizes)}")
-        if compression is not None:
-            create_options.append(f"COMPRESS={compression}")
-        if compression_level is not None:
-            create_options.append(f"ZLEVEL={compression_level}")
-
-        # netCDF driver doesn't support SetIndexingVariable — create
-        # dimension arrays manually without linking them.
-        use_set_indexing = driver_type == "MEM"
-
-        # Determine if CRS is geographic (lon/lat) or projected (m)
-        is_geographic = True
-        if epsg is not None:
-            srs_check = sr_from_epsg(int(epsg))
-            is_geographic = srs_check.IsGeographic() == 1
-
-        dim_x = NetCDF._create_dimension(
-            rg,
-            "x",
-            dtype,
-            np.array(x_dim_values),
-            gdal.DIM_TYPE_HORIZONTAL_X,
-            use_set_indexing,
-            is_geographic=is_geographic,
-        )
-        dim_y = NetCDF._create_dimension(
-            rg,
-            "y",
-            dtype,
-            np.array(y_dim_values),
-            gdal.DIM_TYPE_HORIZONTAL_Y,
-            use_set_indexing,
-            is_geographic=is_geographic,
-        )
-
-        # Build one GDAL dimension per non-spatial axis in storage
-        # order, then create the MDArray with `(*extra_dims, dim_y,
-        # dim_x)`. This generalises the previous 3-D-only branch: a
-        # 2-D array yields zero extra dims; a 3-D array yields one;
-        # 4-D+ yields the full set. The first non-spatial dim is
-        # tagged `DIM_TYPE_TEMPORAL` (matching the legacy 3-D path),
-        # and any additional dims are left untagged so the netCDF
-        # driver doesn't second-guess their semantics.
-        gdal_extra_dims = []
-        for i, (dim_name, dim_values) in enumerate(extra_dims):
-            dim_type = gdal.DIM_TYPE_TEMPORAL if i == 0 else None
-            gd_dim = NetCDF._create_dimension(
-                rg,
-                dim_name,
-                dtype,
-                np.array(dim_values),
-                dim_type,
-                use_set_indexing,
-            )
-            gdal_extra_dims.append(gd_dim)
-        md_arr = rg.CreateMDArray(
-            variable_name,
-            [*gdal_extra_dims, dim_y, dim_x],
-            dtype,
-            create_options if create_options else [],
-        )
-
-        # Set metadata BEFORE writing data — netCDF driver requires
-        # nodata to be set before the first Write call.
-        # Tolerate both scalar and per-band sequence inputs since
-        # callers often pass `Dataset.no_data_value` (now a tuple)
-        # straight through.
-        ndv_scalar = (
-            no_data_value[0]
-            if isinstance(no_data_value, (list, tuple)) and no_data_value
-            else no_data_value
-        )
-        if ndv_scalar is not None:
-            md_arr.SetNoDataValueDouble(float(ndv_scalar))
-        if epsg is None:
-            raise ValueError("epsg cannot be None")
-        srse = sr_from_epsg(int(epsg))
-        md_arr.SetSpatialRef(srse)
-        md_arr.Write(arr)
-
-        # Create CF grid_mapping variable (MEM driver only — the netCDF
-        # driver creates its own via SetSpatialRef above). Use
-        # "spatial_ref" as the variable name to avoid collision with
-        # GDAL's automatic "crs" during CreateCopy to netCDF.
-        if driver_type == "MEM":
-            gm_name, gm_params = srs_to_grid_mapping(srse)
-            gm_dtype = gdal.ExtendedDataType.Create(gdal.GDT_Int32)
-            gm_var_name = "spatial_ref"
-            crs_arr = rg.CreateMDArray(gm_var_name, [], gm_dtype)
-            crs_arr.Write(np.array(0, dtype=np.int32))
-            gm_params["grid_mapping_name"] = gm_name
-            write_attributes_to_md_array(crs_arr, gm_params)
-            write_attributes_to_md_array(md_arr, {"grid_mapping": gm_var_name})
-
-        return src
+        return _variables.create_from_array(*args, **kwargs)
 
     @staticmethod
     def _resolve_dst_dimensions(dst_group, src_dims):
